feat: Option for enabling auto-hiding scrollbars, documentation
