+++ README.md
@@ -22,13 +22,14 @@ pip install -e .
 # Documentation
 
 ### `CodeView` Widget
-|Options      |Description                     |Input                                         |
-|-------------|--------------------------------|----------------------------------------------|
-|master       |The parent widget               |Tkinter widget                                |
-|lexer        |The Language lexer              |Pygments lexer                                |
-|color_scheme |A color scheme for the code     |Dict, string, or toml file                    |
-|tab_width    |The width of a tab (`\t`)       |Int                                           |
-|**kwargs     |Keyword arguments for the widget|Any keyword arguments given to a `Text` widget|
+|Options           |Description                     |Input                                         |
+|------------------|--------------------------------|----------------------------------------------|
+|master            |The parent widget               |Tkinter widget                                |
+|lexer             |The Language lexer              |Pygments lexer                                |
+|color_scheme      |A color scheme for the code     |Dict, string, or toml file                    |
+|tab_width         |The width of a tab (`\t`)       |Int                                           |
+|autohide_scrollbar|Auto hide scrollbars            |Bool                                           |
+|**kwargs          |Keyword arguments for the widget|Any keyword arguments given to a `Text` widget|
 
 #### Basic Usage:
 ```python
